Make all extensions public (move Color.hsl extension to public)
